+++ Sources/TuistScaffoldTesting/Utils/Mocks/MockTemplatesDirectoryLocator.swift
@@ -4,18 +4,23 @@ import TSCBasic
 
 public final class MockTemplatesDirectoryLocator: TemplatesDirectoryLocating {
     public var locateUserTemplatesStub: ((AbsolutePath) -> AbsolutePath?)?
-    public var templateDirectoriesStub: ((AbsolutePath) throws -> [AbsolutePath])?
     public var locateTuistTemplatesStub: (() -> AbsolutePath?)?
+    public var templateDirectoriesStub: ((AbsolutePath) throws -> [AbsolutePath])?
+    public var templatePluginDirectoriesStub: ((AbsolutePath) throws -> [AbsolutePath])?
 
     public func locateUserTemplates(at: AbsolutePath) -> AbsolutePath? {
         locateUserTemplatesStub?(at)
+    }
+
+    public func locateTuistTemplates() -> AbsolutePath? {
+        locateTuistTemplatesStub?()
     }
 
     public func templateDirectories(at path: AbsolutePath) throws -> [AbsolutePath] {
         try templateDirectoriesStub?(path) ?? []
     }
 
-    public func locateTuistTemplates() -> AbsolutePath? {
-        locateTuistTemplatesStub?()
+    public func templatePluginDirectories(at path: AbsolutePath) throws -> [AbsolutePath] {
+        try templatePluginDirectoriesStub?(path) ?? []
     }
 }