+++ fixtures/app_with_plugins/Tuist/Config.swift
@@ -3,7 +3,7 @@ import ProjectDescription
  let config = Config(
      plugins: [
          .local(path: .relativeToManifest("../../LocalPlugin")),
-         .git(url: "https://github.com/tuist/ExampleTuistPlugin.git", tag: "1.0.0")
+         .git(url: "https://github.com/tuist/ExampleTuistPlugin.git", tag: "2.0.0")
      ],
      generationOptions: []
  )